+++ source/core/PD777.cpp
@@ -2052,6 +2052,8 @@ PD777::execJS(const u16 pc, const u16 code)
 void
 PD777::init()
 {
+    config.setDefault();
+    setupConfig(config);
     setupCode();
     setupPattern();
     crt.reset();